storage fixes

+++ server/src/main/scala/pl/touk/wsr/server/storage/InMemoryStorageWithSerialization.scala
@@ -9,7 +9,10 @@ import pl.touk.wsr.server.utils.ScalaUtils._
 import scala.concurrent.Future
 import scala.util.{Failure, Success, Try}
 
-class InMemoryStorageWithSerialization(dataPackSize: Int, maxPacksDataSize: Int, dataPath: String)
+class InMemoryStorageWithSerialization(dataPackSize: Int,
+                                       maxPacksDataSize: Int,
+                                       dataPath: String,
+                                       serializable: Boolean = true)
   extends Storage with LazyLogging {
 
   private case class UUIDDataPackId(uuid: UUID) extends DataPackId
@@ -31,18 +34,18 @@ class InMemoryStorageWithSerialization(dataPackSize: Int, maxPacksDataSize: Int,
 
   override def addData(number: Int): Future[Unit] = Future.successful {
     synchronized {
+      logStorageState("adding")
       val newUnpackedData = unpackedData :+ number
       if (newUnpackedData.length == dataPackSize) {
-        logStorageState("adding")
         val dataPack = DataPackWithReservation(DataPack(UUIDDataPackId(), newUnpackedData))
         dataPacks = dataPacks :+ dataPack
         unpackedData = List.empty[Int]
-        waitingFor = waitingFor - dataPackSize
-        logStorageState("after adding")
       } else {
         unpackedData = newUnpackedData
       }
+      waitingFor = waitingFor -1
       save()
+      logStorageState("after adding")
     }
   }
 
@@ -91,12 +94,11 @@ class InMemoryStorageWithSerialization(dataPackSize: Int, maxPacksDataSize: Int,
   }
 
   private def logStorageState(info: String) = {
-    val maxSize = maxPacksDataSize * dataPackSize
     logger.debug(s"""
         |---------  STORAGE ---------
         |--------- $info
-        | max-size:  $maxSize
-        | free-size: ${maxSize - dataPacks.size * dataPackSize}
+        | max-size: ${maxPacksDataSize * dataPackSize}
+        | use-size: ${dataPacks.size * dataPackSize}
         |----------------------------
         |""".stripMargin
     )
@@ -104,34 +106,42 @@ class InMemoryStorageWithSerialization(dataPackSize: Int, maxPacksDataSize: Int,
 
   override def freeRequestedDataSpace: Future[Unit] = Future.successful {
     synchronized {
-      // todo: maybe: nextOffset = nextOffset - waitingFor
+      nextOffset = nextOffset - waitingFor
       waitingFor = 0
     }
   }
 
-  private def save() = Try {
-    val oos = new ObjectOutputStream(new FileOutputStream(dataPath))
-    oos.writeObject(dataPacks.flatMap(_.dataPack.sequence))
-    oos.close()
-  } match {
-    case r@Success(_) =>
-      r
-    case r@Failure(ex) =>
-      logger.error("Serialization fails!", ex)
-      Try(new File(dataPath).delete())
-      r
+  private def save() = {
+    if(serializable) {
+      Try {
+        val oos = new ObjectOutputStream(new FileOutputStream(dataPath))
+        oos.writeObject(dataPacks.flatMap(_.dataPack.sequence))
+        oos.close()
+      } match {
+        case r@Success(_) =>
+          r
+        case r@Failure(ex) =>
+          logger.error("Serialization fails!", ex)
+          Try(new File(dataPath).delete())
+          r
+      }
+    }
   }
 
   private def load(): Try[Seq[DataPackWithReservation]] = {
-    Try {
-      val ois = new ObjectInputStream(new FileInputStream(dataPath))
-      val loadedDataPacks = ois.readObject.asInstanceOf[Seq[Int]]
-      ois.close()
-      loadedDataPacks
-    } map { list =>
-      list.grouped(dataPackSize).map(group =>
-        DataPackWithReservation(DataPack(UUIDDataPackId(), group), reserved = false)
-      ).toSeq
+    if(serializable) {
+      Try {
+        val ois = new ObjectInputStream(new FileInputStream(dataPath))
+        val loadedDataPacks = ois.readObject.asInstanceOf[Seq[Int]]
+        ois.close()
+        loadedDataPacks
+      } map { list =>
+        list.grouped(dataPackSize).map(group =>
+          DataPackWithReservation(DataPack(UUIDDataPackId(), group), reserved = false)
+        ).toSeq
+      }
+    } else {
+      Failure(new Exception("disabled"))
     }
   }
